requirements.txt
allure-behave==2.13.2
allure-python-commons==2.13.2
attrs==23.2.0
behave==1.2.6
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
exceptiongroup==1.2.0
h11==0.14.0
idna==3.6
iniconfig==2.0.0
outcome==1.3.0.post0
packaging==24.0
parse==1.20.1
parse-type==0.6.2
pluggy==1.4.0
py==1.11.0
pycparser==2.22
PySocks==1.7.1
pytest==7.4.0
PyYAML==6.0.2
requests==2.31.0
selenium==4.11.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
tomli==2.0.1
trio==0.25.0
trio-websocket==0.11.1
Unidecode==1.3.6
urllib3==2.2.1
wsproto==1.2.0
